third_party/WebKit/LayoutTests/http/tests/inspector/network/long-script-content.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script src="../debugger-test.js"></script>
<script src="../../../../resources/gc.js"></script>
<script>

var scriptElement;
function loadScript()
{
    scriptElement = document.createElement("script");
    scriptElement.src = "resources/long_script.cgi";
    document.head.appendChild(scriptElement);
}

function unloadScript()
{
    scriptElement.parentElement.removeChild(scriptElement);
}

function test() {
  TestRunner.hardReloadPage(step1);

  function step1() {
    ConsoleTestRunner.addConsoleSniffer(step2);
    TestRunner.evaluateInPage('loadScript()');
  }

  function step2(event) {
    TestRunner.evaluateInPage('unloadScript()', step3);
  }

  function step3() {
    SourcesTestRunner.waitForScriptSource('long_script.cgi', step4);
  }

  function step4(uiSourceCode) {
    TestRunner.evaluateInPage('gc()', step5.bind(null, uiSourceCode));
  }

  function step5(uiSourceCode) {
    uiSourceCode.requestContent().then(step6);
  }

  function step6(loadedScript) {
    var expected = 'console.log(\'finished\');\n';
    TestRunner.assertTrue(!!loadedScript, 'No script content');
    loadedScript = loadedScript.replace(/\r\n/g, '\n');  // on windows we receive additional symbol \r at line end.
    TestRunner.assertEquals(1024 * 10240 + expected.length, loadedScript.length, 'Loaded script length mismatch');
    var actual = loadedScript.substring(loadedScript.length - expected.length);
    TestRunner.assertEquals(expected, actual, 'Loaded script is corrupted');

    TestRunner.completeTest();
  }
}
</script>
</head>
<body onload="runTest()">
    <p> Tests long script content is correctly shown in source panel after page reload.</p>
</body>
</html>
